+ Rename sfdx to sf as the latest client version
